+++ CHANGELOG.md
@@ -3,10 +3,13 @@
 ### 1.6.6
 - Added new Crafter modes: ignore redstone signal, redstone signal unlocks autocrafting, redstone signal locks autocrafting and redstone pulse inserts next set (replacement for blocking mode) (raoulvdberge)
 - Added a config option to configure the autocrafting calculation timeout in milliseconds (raoulvdberge)
+- Added throttling for network devices that can request autocrafting (raoulvdberge)
 - Fixed an autocrafting bug where it crashed when external inventories couldn't be filled (raoulvdberge)
 - Fixed a duplication bug with a disconnected Crafting Grid (raoulvdberge)
 - Fixed oredict autocrafting sometimes reporting that a craftable item is missing (raoulvdberge)
+- Fixed fluid autocrafting without item inputs locking when there's not enough space for the fluids (raoulvdberge)
 - Removed handling of reusable items in autocrafting, to avoid problems (raoulvdberge)
+- You can no longer start a crafting task if it has missing items or fluids (raoulvdberge)
 
 ### 1.6.5
 - Fixed Refined Storage silicon's oredict entry being registered too late (raoulvdberge)